requirements.txt
aiohttp>=3.13.0
Pillow>=12.0.0
dnspython>=2.8.0
websockets>=16.0.0
